// frontend_restaurants/web/src/lib/types.ts
export type orderItems = orderItem[];

type OrderType = {
	orderId: number;
	type: string;
	orderTime: Date;
	orderItems: orderItems;
	noteFromUser: string;
	totalPrice: number;
	finishTime: Date;
	completeTime: Date;
	rating: number;
	comment: string;
};

export type IncomingOrderType = Omit<
	OrderType,
	'finishTime' | 'completeTime' | 'rating' | 'comment'
>;
export type ProcessingOrderType = Omit<
	OrderType,
	'finishTime' | 'completeTime' | 'rating' | 'comment'
>;
export type WaitingOrderType = Omit<OrderType, 'completeTime' | 'rating' | 'comment'>;
export type CompletedOrderType = Omit<OrderType, 'rating' | 'comment'>;
export type HistoryOrderType = Omit<OrderType, 'completeTime' | 'finishTime' | 'noteFromUser'>;

export type OrderList = OrderType[];
export type IncomingOrderList = IncomingOrderType[];
export type ProcessingOrderList = ProcessingOrderType[];
export type WaitingOrderList = WaitingOrderType[];
export type CompletedOrderList = CompletedOrderType[];
export type HistoryOrderList = HistoryOrderType[];
// ...
export type HistoryRating = number[];